+++ src/utils/misc/helpers.ts
@@ -1,18 +1,3 @@
-/**
- * Return number in formatted number string.
- *
- * ```ts
- * console.log(formatNumber(12345678)); // 12,345,678
- * ```
- */
-export function formatNumber(num: number): string {
-   return num
-      .toString()
-      .split('.')
-      .map((str, i) => (i ? str : str.replace(/\B(?=(\d{3})+(?!\d))/g, ',')))
-      .join('.');
-}
-
 /** Check if string is valid hexadecimal. */
 export function isHex(hex: string): boolean {
    return /^[a-fA-F0-9]+$/g.test(hex);